+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added `Primitives::ToEnumerable` method (and relative extension method).
+
 ## [1.1.2] - 2021-12-19
 
 - Updated `Strings::ContainsSpecialChar`.